Fix CATEGORY to be based on INFFILE, not FIRMWARE
Previously, CATEGORY wasn't getting set correctly because it was
looking in the directory name for "UEFI" or "ME", but the directory
names (at least for the Surface Laptop Go) are often lowercase.

Since lowercase "me" might match too many other directories (e.g.,
surfacesystemtelemetry), I've changed it to check the name of the .inf
file which (at least in the examples I've seen) always have capital
letters for SurfaceME.inf and SurfaceUEFI.inf.